requirements.txt
python-telegram-bot==21.6
asyncpg==0.30.0
anthropic>=0.49.0
python-dotenv==1.0.1
httpx>=0.27.0
PlexAPI>=4.15.0
aiohttp>=3.9.0